+++ src/pages/PostsManagerPage.tsx
@@ -3,7 +3,7 @@ import { Button, Card, CardContent, CardHeader, CardTitle } from "../shared/ui"
 import { URLParams } from "../shared/types"
 import { useTag } from "../features/model/tag/useTag"
 import { usePost } from "../shared/model/usePost"
-import { useURLParams } from "../shared/model/useURLParams"
+import { useURLParams } from "../features/model/url/useURLParams"
 import { PostSearchInput } from "../widgets/ui/post/PostSearchInput"
 import { PostTagSelect } from "../widgets/ui/post/PostTagSelect"
 import { PostSortSelect } from "../widgets/ui/post/PostSortSelect"
